+++ litebox_broker_userland/src/lib.rs
@@ -1,12 +1,24 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-//! Support for the Linux-userland broker process.
+//! Library support for the userland broker process.
 //!
-//! The broker executable composes `litebox_broker_core` and
-//! `litebox_broker_host` with the Linux host endpoints from
-//! `litebox_broker_transport_linux_userland`. It owns runner process lifecycle,
-//! socket setup, and worker threads. This library exposes components shared by
-//! the executable and its integration tests, currently readiness publication.
+//! This crate owns the platform-independent parts of hosting a broker: a
+//! structured builder that selects platform providers and constructs a
+//! [`litebox_broker_core::BrokerCore`] ([`builder`]), the generic hosted
+//! association runtime that serves one association from setup through
+//! teardown ([`runtime`]), and threaded readiness publication
+//! ([`readiness`]). The `litebox-broker-userland` binary composes these with
+//! CLI parsing, runner process lifecycle, and the platform-specific transport
+//! endpoints from `litebox_broker_transport_linux_userland` and
+//! `litebox_broker_transport_windows_userland`; an in-process caller can use
+//! the same builder and runtime directly instead.
 
+pub mod builder;
 pub mod readiness;
+pub mod runtime;
+
+mod random;
+mod stdio;
+
+const WORKER_COUNT: usize = 8;
