+++ litebox_broker_userland/src/windows.rs
@@ -5,61 +5,29 @@
 
 use std::error::Error;
 use std::ffi::OsString;
-use std::io::Result as IoResult;
 use std::os::windows::io::AsRawHandle;
 use std::process::Child;
-use std::sync::Arc;
 use std::time::Instant;
 
-use litebox_broker_core::socket::UnsupportedSocketProvider;
 use litebox_broker_core::{BrokerCore, ObjectRights, PolicyEngine};
-use litebox_broker_protocol::message::{BrokerRequest, BrokerResponse};
 use litebox_broker_protocol::shared_buffer::SHARED_BUFFER_POOL_SIZE;
-use litebox_broker_transport::channel::HostReceive;
-use litebox_broker_transport_windows_userland::control_ring::{
-    WindowsControlRingHostRequestSource, WindowsControlRingHostResponseSink,
-    WindowsControlRingHostShutdown,
-};
 use litebox_broker_transport_windows_userland::named_pipe::{
     WindowsNamedPipeHostSetupChannel, WindowsNamedPipeListener, validate_client_process,
 };
 use litebox_broker_transport_windows_userland::shared_memory::WindowsSharedMemory;
+use litebox_broker_userland::builder::HostedBrokerBuilder;
 
-use super::{
-    HostAssociationShutdown, HostRequestSource, HostResponseSink, SETUP_TIMEOUT,
-    configured_socket_policy,
-};
-use super::{random::UserlandRandomProvider, stdio::UserlandStdioProvider};
-
-impl HostRequestSource for WindowsControlRingHostRequestSource {
-    fn recv_request(&mut self) -> IoResult<HostReceive<BrokerRequest>> {
-        Self::recv_request(self)
-    }
-}
-
-impl HostResponseSink for WindowsControlRingHostResponseSink {
-    fn send_response(&self, response: &BrokerResponse) -> IoResult<()> {
-        Self::send_response(self, response)
-    }
-}
-
-impl HostAssociationShutdown for WindowsControlRingHostShutdown {
-    fn shutdown(&self) -> IoResult<()> {
-        Self::shutdown(self)
-    }
-}
+use super::{SETUP_TIMEOUT, configured_socket_policy};
 
 pub(super) fn run(args: super::CliArgs) -> Result<(), Box<dyn Error>> {
     let control_pipe = unique_control_pipe_name();
     let control_listener = WindowsNamedPipeListener::bind(&control_pipe)?;
-    let broker = BrokerCore::new(
+    let broker = HostedBrokerBuilder::new(
         PolicyEngine::with_host_guaranteed_rights(ObjectRights::all()).with_socket_policy(
             configured_socket_policy(&args.allow_tcp_destination, &args.allow_udp_destination)?,
         ),
-        Arc::new(UnsupportedSocketProvider),
-        Arc::new(UserlandRandomProvider),
-        Arc::new(UserlandStdioProvider::new()?),
-    )?;
+    )
+    .build()?;
 
     crate::run_runner_process(&args, &control_pipe, None, |runner, runner_process_id| {
         serve_runner(&broker, control_listener, runner, runner_process_id)
@@ -80,7 +48,7 @@ fn serve_runner(
     let control_channel =
         WindowsNamedPipeHostSetupChannel::from_host_guaranteed(control_stream, setup_deadline);
     let runner_process = runner.as_raw_handle();
-    crate::serve_runner(
+    litebox_broker_userland::runtime::serve_association(
         broker,
         control_channel,
         || WindowsSharedMemory::create(SHARED_BUFFER_POOL_SIZE),
@@ -90,7 +58,8 @@ fn serve_runner(
             channel.send_shared_memory(control_memory, runner_process)
         },
         WindowsNamedPipeHostSetupChannel::into_active,
-    )
+    )?;
+    Ok(())
 }
 
 fn unique_control_pipe_name() -> OsString {
